refactor(user_service): move UserServiceMapper to application/mapper
