+++ src/main/java/com/renewsim/backend/user_service/infrastructure/persistence/adapter/UserPersistenceAdapter.java
@@ -6,7 +6,7 @@ import com.renewsim.backend.user_service.application.port.out.RoleCatalogPort;
 import com.renewsim.backend.user_service.application.port.out.UserRepositoryPort;
 import com.renewsim.backend.user_service.domain.model.User;
 import com.renewsim.backend.user_service.dto.UserFilterRequest;
-import com.renewsim.backend.user_service.infrastructure.mapper.UserServiceMapper;
+import com.renewsim.backend.user_service.application.mapper.UserServiceMapper;
 import com.renewsim.backend.user_service.infrastructure.persistence.entity.UserEntity;
 import com.renewsim.backend.user_service.infrastructure.persistence.repo.UserJpaRepository;
 import lombok.RequiredArgsConstructor;
