+++ src/test/java/com/renewsim/backend/user_service/infrastructure/mapper/UserMapperTest.java
@@ -2,6 +2,7 @@ package com.renewsim.backend.user_service.infrastructure.mapper;
 
 import com.renewsim.backend.shared.domain.vo.RoleName;
 import com.renewsim.backend.role_service.infrastructure.persistence.entity.RoleEntity;
+import com.renewsim.backend.user_service.application.mapper.UserServiceMapper;
 import com.renewsim.backend.user_service.domain.model.User;
 import com.renewsim.backend.user_service.dto.UserCreateRequest;
 import com.renewsim.backend.user_service.dto.UserResponse;
